+++ backend/danswer/connectors/gmail/connector.py
@@ -9,24 +9,16 @@ from googleapiclient.discovery import Resource  # type: ignore
 
 from danswer.configs.app_configs import INDEX_BATCH_SIZE
 from danswer.configs.constants import DocumentSource
-from danswer.connectors.cross_connector_utils.google.google_auth import get_google_creds
-from danswer.connectors.cross_connector_utils.google.google_utils import (
-    execute_paginated_retrieval,
-)
-from danswer.connectors.cross_connector_utils.google.shared_constants import (
+from danswer.connectors.cross_connector_utils.miscellaneous_utils import time_str_to_utc
+from danswer.connectors.google_utils.google_auth import get_google_creds
+from danswer.connectors.google_utils.google_utils import execute_paginated_retrieval
+from danswer.connectors.google_utils.shared_constants import (
     DB_CREDENTIALS_PRIMARY_ADMIN_KEY,
 )
-from danswer.connectors.cross_connector_utils.google.shared_constants import (
-    MISSING_SCOPES_ERROR_STR,
-)
-from danswer.connectors.cross_connector_utils.google.shared_constants import (
-    ONYX_SCOPE_INSTRUCTIONS,
-)
-from danswer.connectors.cross_connector_utils.google.shared_constants import (
-    SLIM_BATCH_SIZE,
-)
-from danswer.connectors.cross_connector_utils.google.shared_constants import USER_FIELDS
-from danswer.connectors.cross_connector_utils.miscellaneous_utils import time_str_to_utc
+from danswer.connectors.google_utils.shared_constants import MISSING_SCOPES_ERROR_STR
+from danswer.connectors.google_utils.shared_constants import ONYX_SCOPE_INSTRUCTIONS
+from danswer.connectors.google_utils.shared_constants import SLIM_BATCH_SIZE
+from danswer.connectors.google_utils.shared_constants import USER_FIELDS
 from danswer.connectors.interfaces import GenerateDocumentsOutput
 from danswer.connectors.interfaces import GenerateSlimDocumentOutput
 from danswer.connectors.interfaces import LoadConnector
