+++ backend/danswer/connectors/google_drive/connector.py
@@ -6,26 +6,6 @@ from google.oauth2.service_account import Credentials as ServiceAccountCredentia
 
 from danswer.configs.app_configs import INDEX_BATCH_SIZE
 from danswer.configs.constants import DocumentSource
-from danswer.connectors.cross_connector_utils.google.google_auth import get_google_creds
-from danswer.connectors.cross_connector_utils.google.google_utils import (
-    execute_paginated_retrieval,
-)
-from danswer.connectors.cross_connector_utils.google.shared_constants import (
-    DB_CREDENTIALS_PRIMARY_ADMIN_KEY,
-)
-from danswer.connectors.cross_connector_utils.google.shared_constants import (
-    MISSING_SCOPES_ERROR_STR,
-)
-from danswer.connectors.cross_connector_utils.google.shared_constants import (
-    ONYX_SCOPE_INSTRUCTIONS,
-)
-from danswer.connectors.cross_connector_utils.google.shared_constants import (
-    SCOPE_DOC_URL,
-)
-from danswer.connectors.cross_connector_utils.google.shared_constants import (
-    SLIM_BATCH_SIZE,
-)
-from danswer.connectors.cross_connector_utils.google.shared_constants import USER_FIELDS
 from danswer.connectors.google_drive.doc_conversion import (
     convert_drive_item_to_document,
 )
@@ -36,6 +16,16 @@ from danswer.connectors.google_drive.models import GoogleDriveFileType
 from danswer.connectors.google_drive.resources import get_admin_service
 from danswer.connectors.google_drive.resources import get_drive_service
 from danswer.connectors.google_drive.resources import get_google_docs_service
+from danswer.connectors.google_utils.google_auth import get_google_creds
+from danswer.connectors.google_utils.google_utils import execute_paginated_retrieval
+from danswer.connectors.google_utils.shared_constants import (
+    DB_CREDENTIALS_PRIMARY_ADMIN_KEY,
+)
+from danswer.connectors.google_utils.shared_constants import MISSING_SCOPES_ERROR_STR
+from danswer.connectors.google_utils.shared_constants import ONYX_SCOPE_INSTRUCTIONS
+from danswer.connectors.google_utils.shared_constants import SCOPE_DOC_URL
+from danswer.connectors.google_utils.shared_constants import SLIM_BATCH_SIZE
+from danswer.connectors.google_utils.shared_constants import USER_FIELDS
 from danswer.connectors.interfaces import GenerateDocumentsOutput
 from danswer.connectors.interfaces import GenerateSlimDocumentOutput
 from danswer.connectors.interfaces import LoadConnector
